% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_subsets.R
\name{all_subsets}
\alias{all_subsets}
\title{all_subsets Function}
\usage{
all_subsets(series, covariates, min, max, type, fit = TRUE)
}
\arguments{
\item{series}{Data with annual values of response (abundance) and predictor covariates.}

\item{covariates}{Vector of covariates to include in models.}

\item{min}{Minimum number of covariates within individual model.}

\item{max}{Maximum number of covariates within individual model.}

\item{type}{Specifies whether an in-season or pre-season forecast.}

\item{fit}{Flag for whether to fit models.}
}
\value{
A list containing ...

all_subsets Function
}
\description{
This function performs ...
}
\examples{
\dontrun{
  library(allsubsetspackagenew)  # Load your package

  # Example code for using the all_subsets function
  data_file <- system.file("data", "dat.csv", package = "allsubsetspackagenew")
  data <- read.csv(data_file)  # Load your data

  covariates <- c(
    "lag1_log_JackOPI",
    "lag1_log_SmAdj",
    "lag1_NPGO",
    "lag1_PDO",
    "WSST_A",
    "PDO.MJJ",
    "MEI.OND",
    "UWI.JAS",
    "SST.AMJ",
    "SSH.AMJ",
    "UWI.SON"
  )

  min_covariates <- 0
  max_covariates <- 1
  forecast_type <- "preseason"

  results <- all_subsets(
    series = data,
    covariates = covariates,
    min = min_covariates,
    max = max_covariates,
    type = forecast_type,
    fit = TRUE
  )

  cat("Number of models:", length(results[[1]]), "\n")
  cat("Best model formula:", results[[2]]$formula[1], "\n")
  cat("AICc of the best model:", results[[2]]$AICc[1], "\n")
}


}
